requirements.txt
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
paramiko==3.4.1
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
